Move dashboard_activity out of pages/
