+++ application/plotlydash/pages/dashboard_db.py
@@ -13,7 +13,7 @@ import plotly.graph_objs as go
 
 from application import converters, stravatalk, util
 from application.models import db, Activity
-from application.plotlydash.pages import dashboard_activity
+from application.plotlydash import dashboard_activity
 # calc_power, create_rows, init_hover_callbacks
